Socket events and functions are seperated from server file and moved to socketEvents.js file to increase understandibility and maintibility, worked on code formatting
